+++ braintree/amex_express_checkout_card.py
@@ -1,11 +1,13 @@
 import braintree
 from braintree.resource import Resource
+from warnings import warn
 
 class AmexExpressCheckoutCard(Resource):
     """
-    A class representing Braintree Amex Express Checkout card objects.
+    A class representing Braintree Amex Express Checkout card objects. Deprecated
     """
     def __init__(self, gateway, attributes):
+        warn("AmexExpressCheckoutCard is deprecated")
         Resource.__init__(self, gateway, attributes)
 
         if "subscriptions" in attributes:
@@ -14,4 +16,3 @@ class AmexExpressCheckoutCard(Resource):
     @property
     def expiration_date(self):
         return self.expiration_month + "/" + self.expiration_year
-