+++ braintree/masterpass_card.py
@@ -1,12 +1,14 @@
 import braintree
 from braintree.address import Address
 from braintree.resource import Resource
+from warnings import warn
 
 class MasterpassCard(Resource):
     """
-    A class representing Masterpass card
+    A class representing Masterpass card. Deprecated
     """
     def __init__(self, gateway, attributes):
+        warn("MasterpassCard is deprecated")
         Resource.__init__(self, gateway, attributes)
 
         if "billing_address" in attributes: