+++ braintree/payment_instrument_type.py
@@ -6,10 +6,12 @@ class PaymentInstrumentType():
     CreditCard = "credit_card"
     ApplePayCard = "apple_pay_card"
     AndroidPayCard = "android_pay_card"
+    # NEXT_MAJOR_VERSION remove amex express checkout
     AmexExpressCheckoutCard = "amex_express_checkout_card"
     VenmoAccount = "venmo_account"
     UsBankAccount = "us_bank_account"
     VisaCheckoutCard = "visa_checkout_card"
+    # NEXT_MAJOR_VERSION remove masterpass
     MasterpassCard = "masterpass_card"
     SamsungPayCard = "samsung_pay_card"
     LocalPayment = "local_payment"